Changed datatype to u32 (not u64) for increased performance

+++ 2020/src/day15_20.rs
@@ -2,17 +2,17 @@ use super::tools;
 use std::time::Instant;
 use std::collections::BTreeMap;
 
-pub fn get_number(numbers: &Vec<u64>, it: u64) -> u64 {
-    let mut memory: BTreeMap<u64,u64> = BTreeMap::new();
-    let mut last: u64;
+pub fn get_number(numbers: &Vec<u32>, it: u32) -> u32 {
+    let mut memory: BTreeMap<u32,u32> = BTreeMap::new();
+    let mut last: u32;
 
     last = numbers[0];
     for (i, &n) in numbers.iter().skip(1).enumerate() {
-        memory.insert(last,i as u64 + 2);
+        memory.insert(last,i as u32 + 2);
         last = n;
     }
 
-    for count in numbers.len() as u64..it {
+    for count in numbers.len() as u32..it {
         let next;
         let turn = count + 1;
         if let Some(time) = memory.get(&last) {
@@ -43,9 +43,9 @@ pub fn run(real: bool, print_result: bool) -> (u128, u128, u128) {
         "./input/day15_20_real.txt"
     };
     let input = tools::get_input(String::from(input_file));
-    let numbers: Vec<u64> = input[0]
+    let numbers: Vec<u32> = input[0]
         .split(',')
-        .map(|s| s.parse::<u64>().unwrap())
+        .map(|s| s.parse::<u32>().unwrap())
         .collect();
     
     let after0 = Instant::now();
